Move decorators import back to submodule level
