+++ indra/pipeline/__init__.py
@@ -1,1 +1,2 @@
 from .pipeline import AssemblyPipeline, RunnableArgument
+from .decorators import register_pipeline, pipeline_functions
